Rename GenerateServiceAccountDialog component
